feat e refactor:adição do 'editar intervalo sono' em SonoService e refatorações
